requirements.txt
jira
selenium
keyring
jira_dumper